moved loggers into own folder
